+++ io.h
@@ -18,7 +18,7 @@
 
 FILE* OpenFile(char* filePath);
 int CloseFile(FILE* fPtr);
-int ReadConfigFile(FILE* fPtr, poly_s* pf, float* xinf, float* xsup, int* intervals); 
+int ReadConfigFile(FILE* fPtr, poly_s* pf, double* xinf, double* xsup, int* intervals); 
 int ParseLine(poly_s* pf, char* str);
 
 #endif
